Displays all images' URLs in each module
